flight-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
flights
python-dotenv==1.0.1
cachetools==5.5.0
apscheduler==3.10.4
firebase-admin==6.5.0
resend==2.6.0
